working non uniform scale

+++ Assets/Scripts/User Interface/Hand Menu/HM_DeleteSelected.cs
@@ -5,6 +5,8 @@ public class HM_DeleteSelected : HM_Base
     public override void OnClick()
     {
         base.OnClick();
+        if (_deps.selection.SelectionExist == false) return;
+
         _deps.selection.DeleteSelected();
         _deps.handMenu.Show(false);
     }
